API for Instances.jl

Index

Instances.CWLinstanceFactory

Instances.GAinstanceFactory

Instances.MCNDinstanceFactory

Instances.UC_instance

Instances.abstractInstance

Instances.abstractInstanceFactory

Instances.abstractInstanceMCND

Instances.compact_model

Instances.cpuCWLinstanceFactory

Instances.cpuGAinstanceFactory

Instances.cpuInstanceCWL

Instances.cpuInstanceGA

Instances.cpuInstanceMCND

Instances.cpuMCNDinstanceFactory

Instances.decomposed_model

Instances.gpuMCNDinstance

Instances.instanceCWL

Instances.instanceGA

Instances.CR

Instances.CR

Instances.CR

Instances.CR

Instances.LR

Instances.LR

Instances.LR

Instances.LR

Instances.LR

Instances.b

Instances.capacity

Instances.cijk

Instances.constantLagrangianBound

Instances.create_LR

Instances.create_LR_component

Instances.create_data_object

Instances.create_data_object

Instances.create_data_object

Instances.create_data_object

Instances.create_data_object

Instances.destination

Instances.fixed_cost

Instances.generate_GA

Instances.head

Instances.indegree

Instances.indegree_k

Instances.isInKij

Instances.lengthLM

Instances.lengthLM

Instances.lengthLM

Instances.modify_instance

Instances.modify_objective

Instances.modify_objectives

Instances.objective_coefficient_type

Instances.origin

Instances.outdegree

Instances.outdegree_k

Instances.print_dat

Instances.print_dat

Instances.print_dat

Instances.read_dat

Instances.read_dat

Instances.read_dat

Instances.read_dat

Instances.read_modify_dat

Instances.read_modify_dat

Instances.routing_cost

Instances.sizeE

Instances.sizeK

Instances.sizeLM

Instances.sizeLM

Instances.sizeLM

Instances.sizeV

Instances.solve_knapsack

Instances.solve_knapsack_continuous

Instances.tail

Instances.value_LR

Instances.value_LR_a

Instances.volume

Public

Instances.CWLinstanceFactory — Type

Abstract Factory for every instance of the Capacitated Warehose Locaation Problem.

Instances.GAinstanceFactory — Type

Abstract Factory for every instance of the Generalized Assignment Problem.

Instances.MCNDinstanceFactory — Type

Abstract type for the Multi Commodity Network Design Problem.

Instances.abstractInstance — Type

Abstract type for the instances.

Instances.abstractInstanceFactory — Type

Abstract type for the instance factories.

Instances.abstractInstanceMCND — Type

Abstract type for every instance of the Multi Commodity Network Design Problem.

Instances.cpuCWLinstanceFactory — Type

Factory structure used to construct an instance of the Bin Packing Problem.

Instances.cpuGAinstanceFactory — Type

Factory structure used to construct an instance of the Bin Packing Problem.

Instances.cpuInstanceCWL — Type

Structure that describe a cpu instance of the Capacitated Warehose Location Problem.

Fields:

-I: facilities number -J: custumers number -c: costs to satisfy the demand of a costumer with a certain facility -f: costs for open facilities -q: capacity of the facilities -d: demands of the costumers

Instances.cpuInstanceGA — Type

Structure that describe an instance of the Generalized Assignment Problem.

Fields:

I: number of items

J: number of bins

p: profit matrix

w: weights matrix

c: capacities of the bins

Instances.cpuInstanceMCND — Type

Structure to describe a MCND instance in such a way that the sub-problem can be solved using CPU.

Fields:

n : nombre de sommets

edges` : table of arcs: [(tₑ,hₑ)]ₑ

K : table of commodities: [(oᵏ, dᵏ, qᵏ)]ₖ

f : fixed costs

r : routing costs

c : capacities of arcs

Instances.cpuMCNDinstanceFactory — Type

Structure for the cpu instance of the Multi Commodity Network Design Problem.

Instances.gpuMCNDinstance — Type

Structure to describe a MCND instance in such a way that the sub-problem can be solved using GPU.

Fields:

n : nombre de sommets

edges` : table of arcs: [(tₑ,hₑ)]ₑ

K : table of commodities: [(oᵏ, dᵏ, qᵏ)]ₖ

f : fixed costs

r : routing costs

c : capacities of arcs

...: other technical parameters that are used for the GPU-computations of the Lagrangian Sub-Problem and are computed just once.

It is not mandatory to provide these parameters to the constructor as they can be automatically constructed starting from the  ones shared with the cpu version of the instance.

Instances.instanceCWL — Type

Abstract type for every instance of the Capacitated Warehose Locaation Problem.

Instances.instanceGA — Type

Abstract type for every instance of the Generalized Assignment Problem.

Instances.CR — Method

Arguments:

ins: an instance structure of type UC_instance.

Solve the linear relaxation of the provided instance and then, it returns the objective value of the linear relaxation, the dual variables associated to flow constraints, the dual variables associated to capacity constraints, the primal flow variables and the primal design variables.

Instances.CR — Method

Arguments:

ins: an instance structure.

Solve the linear relaxation of the provided instance and then, it returns the objective value of the linear relaxation, the dual variables associated to flow constraints, the dual variables associated to capacity constraints, the primal flow variables and the primal design variables.

Instances.CR — Method

Arguments:

ins: a CWL instance structure.

Solve the linear relaxation of the provided instance and then, it returns the objective value of the linear relaxation, the dual variables associated to packing constraints, the dual variables associated to capacity constraints, the primal assigment variables and the primal pack activation variables.

Instances.CR — Method

Arguments:

ins: a GA instance structure.

Solve the linear relaxation of the provided instance and then, it returns the objective value of the linear relaxation, the dual variables associated to packing constraints, the dual variables associated to capacity constraints, the primal assigment variables and the primal pack activation variables.

Instances.LR — Method

Arguments:

-ins: an instance structure of the Bin packing Problem, -π: a vector of Lagrangian Multipliers,

unsplittable: a boolean that say if true that is the unsplittable version of CWL where we have only binary variables.

This function solves the Lagrangian Knapsack Sub-Problem for the provided instance using the provided Lagrangian multipliers vector.

Instances.LR — Method

Arguments:

-ins: an instance structure of the Bin packing Problem -π: a vector of Lagrangian Multipliers

This function solves the Lagrangian Knapsack Sub-Problem for the provided instance using the provided Lagrangian multipliers vector.

Instances.LR — Method

Arguments:

ins: cpuInstanceMCND of the problem

π: a Lagrangian multipliers Vector

Compute LR. regY is the regularization vector for y and α a multiplicative parameter. It does not requires the solution vector x,y as input and it return them as output.

Instances.LR — Method

Arguments:

- `ins`: cpuInstanceMCND of the problem
- `π`: a Lagrangian multipliers Vector
- (`x`,`y`): two vector of the same size of the (primal) solution vector for the sub-problem, not initalized.
- `demands`: a vector [1:K] with the (unsorted) incexes for the demands
- `e`: index of the arc

Compute LR_{ij}, the value of the Lagrangian problem associated with arc of index e.
It requires the solution vector x,y as input and update them durnig the resolution.

Demands is a permutation of 1:p.

Instances.LR — Method

ins: a gpuMCNDinstance to encode the problem

z: a Lagrangian multipliers Vector

Compute LR using the GPU.

Instances.b — Method

Arguments:

- `ins`: a normalized instance structure.
- `i`: the index of the node.
- `k`: the index of the commodity.

Given a certain node and a certain commodity, this function
returns the volume of the demand for the commodity, if the node 
is the origin, the inverse of the volume if the node is the destination
and zero otherwise.

Instances.capacity — Method

Arguments:

- `ins`: a normalized instance structure.
- `e`: the index of the edge.

Returns the capacity of arc ij (that is j)

Instances.cijk — Method

Arguments:

- `ins`: instance of the problem
- `e`: index of the edge
- `k`: index of the commodity
- `π`: a Lagrangian multipliers Vector

Return the cost associated with demand k in the subproblem of L(π) related to arc e

Instances.constantLagrangianBound — Method

Arguments:

- `ins`: cpuInstanceMCND of the problem
- `π`: a Lagrangian multipliers Vector

Compute the component of the bound that does not depends by the soluyion x,y of the sub-problem,
but only by the lagrangian multipliers vector and the demands data.

Instances.create_data_object — Method

Arguments:

G : number of generators

T : time horizon

C_no_load : no load cost

C_marginal : marginal cost

C_startup : startup cost

P_max_gen : maximum power generation

P_min_gen : minimum power generation

P_ramp_up : up-ramp factor when working

P_ramp_down : down-ramp factor when working

P_startup_ramp : up-ramp factor when startup

P_shutdown_ramp: down-ramp factor when shutdown

T_startup_time : minimum time required to be on when the generator is startup

T_shutdown_time: minimum time required to be off when the generator is shutdown

Pd_power_demend: power demand

`Prreserverequirement: power requirement

model: the JuMP model that describe the instance

Instances.create_data_object — Method

Given the data, creates an instance of the Bin Packing Problem.

Arguments:

_: factory

I: number of facilities

J: number of custumers

-c: costs to satisfy the demand of a costumer with a certain facility -f: costs for open facilities -q: capacity of the facilities -d: demands of the costumers

Instances.create_data_object — Method

Arguments:

_: factory of type cpuGAinstanceFactory

I: number of items

J: number of bins

p: profit matrix

w: weights matrix

c: capacities of the bins

Given the data, creates an instance of the Bin Packing Problem.

Instances.create_data_object — Method

Arguments:

::cpuMCNDinstanceFactory: The standard instance factory formulation.

n: The number of node in the graph that defines the instance.

edges: An array of tuble (size ℤ²ˣᴱ), for each edge 'e' we have a tuple defined as (start(e),end(e))

with the starting end ending node for that arc.

K: An array of triplet (with size ℤᵏˣ³), for each commodity k the associated component is (oᵏ,dᵏ,qᵏ), i.e. the origin, destination and volume of the commodity.

f: The vector of fixed costs ( ∈ ℝᴱ)

r: The vector of routing costs ( ∈ ℝᴷˣᴱ)

c: The vector of capacities ( ∈ ℝᴱ)

It create an instance structure ( for an instance in the standard formulation) with the provided inputs data.

Instances.create_data_object — Method

createdataobject(::gpuMCNDinstanceFactory, n, edges, K, f, r, c)

Arguments:

::gpuMCNDinstanceFactory: The standard instance factory formulation.

n: The number of node in the graph that defines the instance.

edges: An array of tuble (size ℤ²ˣᴱ), for each edge 'e' we have a tuple defined as (start(e),end(e))

with the starting end ending node for that arc.

K: An array of triplet (with size ℤᵏˣ³), for each commodity k the associated component is (oᵏ,dᵏ,qᵏ), i.e. the origin, destination and volume of the commodity.

f: The vector of fixed costs ( ∈ ℝᴱ)

r: The vector of routing costs ( ∈ ℝᴷˣᴱ)

c: The vector of capacities ( ∈ ℝᴱ)

It create an instance structure ( for an instance in the standard formulation) with the provided inputs data.

Instances.destination — Method

Arguments:

- `ins`: a normalized instance structure.
- `indexk`: the index of the commodity.

Returns the destination node of the demand k

Instances.fixed_cost — Method

Arguments:

- `ins`: a normalized instance structure.
- `e`: the index of the edge.

Returns the fixed cost of arc e

Instances.generate_GA — Method

Arguments:

seed: random generator seed.

I: number of items.

J: number of bins.

return an instance with I items and J bins generating the weights as:

wᵢⱼ = 1 - 10⋅ϵᵢⱼ

the profits as:

pᵢⱼ = wᵢⱼ/1000 - 10 ⋅ δᵢⱼ

where ϵᵢⱼ, δᵢⱼ are random values in (0,1).

The capacities are genrated as:

Instances.head — Method

Arguments:

- `ins`: a normalized instance structure.
- `indexe`: the index of the edge.

Returns the head of arc ij (that is j)

Instances.indegree — Method

Arguments:

- `ins`: a normalized instance structure.
- `i`: the index of the node.

Returns the indegree of i

Instances.indegree_k — Method

Arguments:

- `ins`: a normalized instance structure.
- `i`: the index of the node.
- `k`: the index of the commodity.

Returns the indegree of i for a given commodity k

Instances.isInKij — Method

Arguments:

- `ins`: a normalized instance structure.
- `e`: the index of the edge.
- `k`: the index of the commodity.

Returns True if k is in Kij

Instances.lengthLM — Method

Arguments:

ins: instance object, should be a sub-type of instanceGA

returns the length of the Lagrangian multipliers for the instance ins.

Instances.lengthLM — Method

Arguments:

ins: instance object, should be a sub-type of instanceCWL

returns the length of the Lagrangian multipliers for the instance ins.

Instances.lengthLM — Method

Arguments:

ins: instance object, should be a sub-type of instanceGA

returns the length of the Lagrangian multipliers for the instance ins.

Instances.modify_instance — Method

Arguments:

-ins: an instance object cpuInstanceMCND, -seed: the random generation seed -x: a number that allows to increase/decrease the number of demands/commodities in the instance as size(newdemands)=round(x*size(newdemands))

return an instance generated straing from ins and perturbing the routing costs, the volumes, the origins and the destinations of the demands. Furthermore it is possible to change the number of demands with the parameter x.

Instances.objective_coefficient_type — Method

Returns the type of the objective for an instance in the standard formulation, that is Int16.

Instances.origin — Method

Arguments:

- `ins`: a normalized instance structure.
- `indexk`: the index of the commodity.

Returns the origin node of the demand k

Instances.outdegree — Method

Arguments:

- `ins`: a normalized instance structure.
- `i`: the index of the node.

Returns the outdegree of i

Instances.outdegree_k — Method

Arguments:

- `ins`: a normalized instance structure.
- `i`: the index of the node.
- `k`: the index of the commodity.

Returns the outdegree of i for a given commodity k

Instances.print_dat — Method

Arguments:

- `path`: the path to the file where we want print the data
	- `ins`: the instance object that we want print in a file, should be a cpuInstanceMCND

print the instance ins in the file in path.

Instances.print_dat — Method

Arguments:

- `path`: the path to the file where we want print the data
	- `ins`: the instance object that we want print in a file, should be a sub-type of instanceCWL

Instances.print_dat — Method

Arguments:

- `path`: the path to the file where we want print the data
	- `ins`: the instance object that we want print in a file, should be a <: instanceGA

print the instance in a .dat file.

Instances.read_dat — Function

Arguments:

path: a string specifying the path to the file

factory: a factory that allows to return the correct instance type, for this function should be cpuUCinstanceFactory

decomposable: a boolean that say if memorize the problems as decomposed sub-problems.

read a data from the file in path and return an instance assocated to the factory.

Instances.read_dat — Method

Arguments:

path : a String that contains the path to the data file of the instance

factory : a cpuCWLinstanceFactory used only to construct an instance of the proper type ( . <: instanceCWL)

Given the path and the factory of the proper type, this function reads the instance information from the file in the provided path and construct an instance of the Capacitated Warehouse Location problem.

Instances.read_dat — Method

Arguments:

-path: a String that contains the path to the data file of the instance -_: a cpuGAinstanceFactory used only to construct an instance of the proper type ( . <: instanceGA)

Given the path and the factory of the proper type, this function reads the instance information from the file in the provided path and construct an instance of the Bin Packing problem.

Instances.read_dat — Method

Arguments:

-path: a String that contains the path to the data file of the instance -factory: a cpuMCNDinstanceFactory used only to construct an instance of the proper type (cpuInstanceMCND)

Given the path and the factory of the proper type, this function reads the instance information from the file in the provided path and construct an instance of the Bin Packing problem.

Instances.read_modify_dat — Function

Arguments:

path: the path to the file .dat containing the information about the instance that we want modify

factory: the instance factory, supports cpuGAinstanceFactory

seed: the random generation seed

α : a parameter that controll the ratio of demands and capacities

newJ: new number of custumers

returns an instance obtained modifying the original one.

Instances.read_modify_dat — Function

Arguments:

path: the path to the file .dat containing the information about the instance that we want modify

factory: the instance factory, supports cpuCWLinstanceFactory

seed: the random generation seed

α : a parameter that controll the ratio of demands and capacities

newJ: new number of custumers

returns an instance obtained modifying the original one.

Instances.routing_cost — Method

Arguments:

- `ins`: a normalized instance structure.
- `e`: the index of the edge.
- `k`: the index of the commodity.

Returns the routing cost

Instances.sizeE — Method

Arguments:

- `ins`: a normalized instance structure.

Returns the number of arcs

Instances.sizeK — Method

Arguments:

- `ins`: a normalized instance structure.

Returns the number of demands

Instances.sizeLM — Method

Arguments:

ins: instance object, should be a sub-type of instanceGA

returns the size of the Lagrangian multipliers for the instance ins.

Instances.sizeLM — Method

Arguments:

ins: instance object, should be a sub-type of instanceCWL

returns the size of the Lagrangian multipliers for the instance ins.

Instances.sizeLM — Method

Arguments:

ins: instance object, should be a sub-type of instanceGA

returns the size of the Lagrangian multipliers for the instance ins.

Instances.sizeV — Method

Arguments:

- `ins`: a normalized instance structure.

Returns the number of nodes

Instances.solve_knapsack — Method

Arguments:

- `n`: the number of objects 
	- `c`: the costs vector of the objects
	- `w`: the weights vector of the objects
	- `C`: the capacity of the knapsack
	- `x`: the decision variable vector, it should have size n, the solution will be contained in this vector after the call of the function.

Solve a binary knapsack with the provided parameters, save the solution in the vector x and return the objective value.

Instances.solve_knapsack_continuous — Method

Arguments:

- `n`: the number of objects 
	- `c`: the costs vector of the objects
	- `w`: the weights vector of the objects
	- `C`: the capacity of the knapsack
	- `x`: the decision variable vector, it should have size n, the solution will be contained in this vector after the call of the function.

Solve a continuous knapsack with the provided parameters, save the solution in the vector x and return the objective value.

Instances.tail — Method

Arguments:

- `ins`: a normalized instance structure.
- `indexe`: the index of the edge.

Returns the tail of arc ij (that is i)

Instances.value_LR — Method

Arguments:

- `ins`: cpuInstanceMCND of the problem
- `π`: Lagrangian multipliers
- (`x`, `y`): Solution of the Lagrangian problem

Return the value of the Lagrangian sub-problem associated, using the lagrangian multipliers vector π and the primal solution (x,y).

Instances.value_LR_a — Method

Arguments:

- `ins`: cpuInstanceMCND of the problem
- `ia`: index of the arc
- `π`: Lagrangian multipliers
- (`x`, `y`): Solution of the Lagrangian problem

Return the value of the Lagrangian problem associated with arc of index ia,
using the lagrangian multipliers vector π and the primal solution (x,y).

Instances.volume — Method

Arguments:

- `ins`: a normalized instance structure.
- `indexk`: the index of the commodity.

Returns the volume of the demand k

Private

Instances.UC_instance — Type

Structure to describe a UC instance in such a way that the sub-problem can be solved using CPU.

Fields:

G : number of generators

T : time horizon

C_no_load : no load cost

C_marginal : marginal cost

C_startup : startup cost

P_max_gen : maximum power generation

P_min_gen : minimum power generation

P_ramp_up : up-ramp factor when working

P_ramp_down : down-ramp factor when working

P_startup_ramp : up-ramp factor when startup

P_shutdown_ramp: down-ramp factor when shutdown

T_startup_time : minimum time required to be on when the generator is startup

T_shutdown_time: minimum time required to be off when the generator is shutdown

Pd_power_demend: power demand

`Prreserverequirement: power requirement

model: the JuMP model that describe Lagrangian relaxation of the instance

Instances.compact_model — Type

Structure that describe the Lagrangian Relaxation of an UC instance in such a way that the sub-problems are written in a unique JuMP model. This should be used only if the solver associated to the model is automatically capable to exploit the decomposable structure.

Instances.decomposed_model — Type

Structure that describe the Lagrangian Relaxation of an UC instance in such a way that the sub-problems are solved in a decomposed formulation. In other world the Lagrangian Sub-Problem are saved as independent JuMP models that are then solved sequentially, but independently.

Instances.create_LR — Method

Arguments:

ins: an instance structure of type UC_instance.

Construct the Lagrangian Relaxation Sub-Problem for non-decomposed sub-problems.

Instances.create_LR_component — Method

Arguments:

ins: an instance structure of type UC_instance.

g: the index of the generator associated to the sub-problem.

Construct the Lagrangian Relaxation Sub-Problem associated to a given generator g in the case of decomposable sub-problems.

Instances.modify_objective — Method

Arguments:

ins: an instance structure of type UC_instance

y1: the dual variables vector associated to Power Demand Constraints

y2: the dual variables vector associated to Reserve Constraints

For Undecomposed sub-problem. Modify the objective function of the Lagrangian Sub-Problem in the model constained in ins considering y1 and y2 as Lagrangian Mutlipliers vectors.

Instances.modify_objectives — Method

Arguments:

ins: an instance structure of type UC_instance

y1: the dual variables vector associated to Power Demand Constraints

y2: the dual variables vector associated to Reserve Constraints

For Decomposed sub-problem. Modify the objectives functions of the Lagrangian Sub-Problems in the model constained in ins considering y1 and y2 as Lagrangian Mutlipliers vectors.